conversation menu
it’s a hack, but that’s okay for now i think.
* standard now has its own parser (wrapper) with a callback/override
which the menu processing can capture.
* updated the way/when queuing discussing happens with a branch for “is
it the player discussing” — not sure what is actually right, but looks
okay-ish for now. this vaguely matches the threaded conversation’s two
discussion actions ( but rolled up into one )
* minor: moved the timing of the matched callback ( now called OnMatch
) out one level for better control of parser results
* minor: renamed Parse  to ParseInput to reduce search conflict with
options.Parse()
* minor: cleanup of some game/object adapter usage
* minor: reimplemented an Any() test in Play